simulate node module api
